cc26x0: move i2c driver to cc26xx_cc13xx
The registers and functionality on `cc26x0` and `cc26x2_cc13x2`
MCUs is _exactly_ the same.

The pinout for the CC1312 and CC1352 boards is:

- SCL = DIO4
- SDL = DIO5

These are the defaults values and can be found on the quick start
guide (swru535c and swru525e), however another pins can be chosen.

Signed-off-by: Jean Pierre Dudey <jeandudey@hotmail.com>